requirements.txt
python-dotenv==0.21.1
selenium==3.141.0